Add IPython configuration
Add configuration files for IPython.  The default configuration holds
most of the details.  The logger configuration tracks inputs and
outputs.  It utilizes a combination of the IPython logging and the
native Python logging module.  For some reason, IPython insists that the
README file must be in the startup directory.  I don't see a point in
tracking this because it is always the same.  So I added a few lines to
the ignore file.

The next step is to figure out how to report the version information.  I
might just change the stream format for the logging.