Correcting typos in install
So after a first run at using the install script, I found some typos.
The unlink method is part of the main os module not os.path.  I need to
use the function call to now in datatime.  And I need to change all
occurrences of a variable name when I change the instantiation.