Fix PATH example
The commented out portion of the bashrc file demonstrates how to check
if the desired path already exists in the PATH variable.  Essentially,
we need to check both the BIN variable and the PATH variable for the new
path.

    for p in $paths; do
        if ! [[ $(echo "$PATH" | grep -q "$p") ]]; then
            bin=$(path_append "$bin" "$p")
        fi
    done

And looking at it again, I have $1 and $2 swapped in the example.  The
above snippet is the correct order and I don't want to redo the commit
message.  Thus this log will have to do.