Unify path splitting.
The paths script now properly loops over the paths in both zsh and bash.
I also adapted the script to fix OSX "helping" correct the path to bash.
I can't seem to figure out how to split the path in a way that both zsh
and bash can agree on.